backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Mail
gunicorn
python-dotenv
flask-login
Flask-Cors
Werkzeug